🚚 Move some geometry functions to geometry
